Removing TAB characters and avoiding debug warning on curl fail
Setting RETURNTRANSFER to 1 for curl returns the transfer as a string of the return value of curl_exec() instead of outputting (printing) it directly. This makes sure phpBB fails silently rather than throw debug error when curl returns an error. 

Discord also doesn't like tab characters in messages, so this removes those from $message and $footer.